leetcode/src/main/java/com/study/leetcode/linkedlist/MergeTwoSortedLists.java
package com.study.leetcode.linkedlist;

import com.study.leetcode.common.ListNode;

/**
 * 21. Merge Two Sorted Lists
 * Easy
 * https://leetcode.com/problems/merge-two-sorted-lists/description/
 * <p>
 * You are given the heads of two sorted linked lists list1 and list2.
 * <p>
 * Merge the two lists into one sorted list. The list should be made by splicing together the nodes of the first two lists.
 * <p>
 * Return the head of the merged linked list.
 * <p>
 * <p>
 * <p>
 * Example 1:
 * - 1 -> 2 -> 4
 * - 1 -> 3 -> 4
 * => 1 -> 1 -> 2 -> 3 -> 4 -> 4
 * <p>
 * Input: list1 = [1,2,4], list2 = [1,3,4]
 * Output: [1,1,2,3,4,4]
 * <p>
 * Example 2:
 * Input: list1 = [], list2 = []
 * Output: []
 * <p>
 * Example 3:
 * Input: list1 = [], list2 = [0]
 * Output: [0]
 * <p>
 * <p>
 * Constraints:
 * <p>
 * The number of nodes in both lists is in the range [0, 50].
 * -100 <= Node.val <= 100
 * Both list1 and list2 are sorted in non-decreasing order.
 */
public class MergeTwoSortedLists {
    public ListNode mergeTwoLists(ListNode list1, ListNode list2) {
        if (list1 == null) return list2;
        if (list2 == null) return list1;

        ListNode result;
        if (list1.val < list2.val) {
            result = new ListNode(list1.val);
            list1 = list1.next;
        } else {
            result = new ListNode(list2.val);
            list2 = list2.next;
        }


        ListNode current = result;
        while (list1 != null || list2 != null) {
            if (list2 == null) {
                current.next = new ListNode(list1.val);
                list1 = list1.next;
                current = current.next;
                continue;
            } else if (list1 == null) {
                current.next = new ListNode(list2.val);
                list2 = list2.next;
                current = current.next;
                continue;
            }

            if (list1.val < list2.val) {
                current.next = new ListNode(list1.val);
                list1 = list1.next;
            } else {
                current.next = new ListNode(list2.val);
                list2 = list2.next;
            }
            current = current.next;
        }
        return result;
    }

    public ListNode solution1(ListNode list1, ListNode list2) {
        if (list1 != null && list2 != null) {
            if (list1.val < list2.val) {
                list1.next = solution1(list1.next, list2);
                return list1;
            } else {
                list2.next = solution1(list1, list2.next);
                return list2;
            }
        }
        if (list1 == null)
            return list2;
        return list1;
    }

    public ListNode solution2(ListNode list1, ListNode list2) {
        ListNode dummy = new ListNode();
        ListNode current = dummy;

        while (list1 != null && list2 != null) {
            if (list1.val < list2.val) {
                current.next = list1;
                list1 = list1.next;
            } else {
                current.next = list2;
                list2 = list2.next;
            }
            current = current.next;
        }

        current.next = (list1 != null) ? list1 : list2;

        return dummy.next;
    }
}
